requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
aiohttp==3.8.5
    # via
    #   cohere
    #   langchain
    #   langchain-community
    #   openai
aiosignal==1.3.1
    # via aiohttp
annotated-types==0.5.0
    # via pydantic
anyio==3.7.1
    # via
    #   httpx
    #   langchain-core
async-timeout==4.0.3
    # via aiohttp
attrs==23.1.0
    # via aiohttp
backoff==2.2.1
    # via cohere
blinker==1.6.2
    # via flask
boto3==1.28.61
    # via -r requirements.in
botocore==1.31.61
    # via
    #   boto3
    #   s3transfer
build==1.0.3
    # via pip-tools
cachetools==5.3.1
    # via google-auth
certifi==2023.7.22
    # via
    #   elastic-transport
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.2.0
    # via
    #   aiohttp
    #   requests
click==8.1.7
    # via
    #   flask
    #   pip-tools
cohere==4.52
    # via -r requirements.in
dataclasses-json==0.5.14
    # via
    #   langchain
    #   langchain-community
elastic-transport==8.4.0
    # via elasticsearch
elasticsearch==8.12.1
    # via
    #   -r requirements.in
    #   langchain-elasticsearch
exceptiongroup==1.2.0
    # via -r requirements.in
fastavro==1.9.4
    # via cohere
filelock==3.13.1
    # via huggingface-hub
flask==2.3.3
    # via
    #   -r requirements.in
    #   flask-cors
flask-cors==4.0.0
    # via -r requirements.in
frozenlist==1.4.0
    # via
    #   aiohttp
    #   aiosignal
fsspec==2024.2.0
    # via huggingface-hub
google-api-core[grpc]==2.14.0
    # via
    #   google-cloud-aiplatform
    #   google-cloud-bigquery
    #   google-cloud-core
    #   google-cloud-resource-manager
    #   google-cloud-storage
google-auth==2.23.2
    # via
    #   google-api-core
    #   google-cloud-core
    #   google-cloud-storage
google-cloud-aiplatform==1.35.0
    # via -r requirements.in
google-cloud-bigquery==3.13.0
    # via google-cloud-aiplatform
google-cloud-core==2.3.3
    # via
    #   google-cloud-bigquery
    #   google-cloud-storage
google-cloud-resource-manager==1.10.4
    # via google-cloud-aiplatform
google-cloud-storage==2.11.0
    # via google-cloud-aiplatform
google-crc32c==1.5.0
    # via google-resumable-media
google-resumable-media==2.6.0
    # via
    #   google-cloud-bigquery
    #   google-cloud-storage
googleapis-common-protos[grpc]==1.61.0
    # via
    #   google-api-core
    #   grpc-google-iam-v1
    #   grpcio-status
grpc-google-iam-v1==0.12.7
    # via google-cloud-resource-manager
grpcio==1.59.3
    # via
    #   google-api-core
    #   google-cloud-bigquery
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
grpcio-status==1.59.3
    # via
    #   -r requirements.in
    #   google-api-core
h11==0.14.0
    # via httpcore
httpcore==1.0.4
    # via httpx
httpx==0.25.2
    # via mistralai
huggingface-hub==0.21.4
    # via tokenizers
idna==3.4
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
importlib-metadata==6.8.0
    # via
    #   -r requirements.in
    #   cohere
itsdangerous==2.1.2
    # via flask
jinja2==3.1.2
    # via flask
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
jsonpatch==1.33
    # via
    #   langchain
    #   langchain-core
jsonpointer==2.4
    # via jsonpatch
langchain==0.1.9
    # via -r requirements.in
langchain-community==0.0.27
    # via langchain
langchain-core==0.1.30
    # via
    #   langchain
    #   langchain-community
    #   langchain-elasticsearch
    #   langchain-mistralai
langchain-elasticsearch==0.1.0
    # via -r requirements.in
langchain-mistralai==0.0.5
    # via -r requirements.in
langsmith==0.1.10
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
markupsafe==2.1.3
    # via
    #   jinja2
    #   werkzeug
marshmallow==3.20.1
    # via dataclasses-json
mistralai==0.1.3
    # via langchain-mistralai
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.0.0
    # via typing-inspect
numexpr==2.8.5
    # via -r requirements.in
numpy==1.25.2
    # via
    #   langchain
    #   langchain-community
    #   langchain-elasticsearch
    #   numexpr
    #   pandas
    #   pyarrow
    #   shapely
openai==0.27.9
    # via -r requirements.in
orjson==3.9.15
    # via
    #   langsmith
    #   mistralai
packaging==23.2
    # via
    #   build
    #   google-cloud-aiplatform
    #   google-cloud-bigquery
    #   huggingface-hub
    #   langchain-core
    #   marshmallow
pandas==2.2.1
    # via mistralai
pip-tools==7.3.0
    # via -r requirements.in
proto-plus==1.22.3
    # via
    #   google-cloud-aiplatform
    #   google-cloud-bigquery
    #   google-cloud-resource-manager
protobuf==4.25.1
    # via
    #   google-api-core
    #   google-cloud-aiplatform
    #   google-cloud-bigquery
    #   google-cloud-resource-manager
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
    #   proto-plus
pyarrow==15.0.1
    # via mistralai
pyasn1==0.5.0
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.3.0
    # via google-auth
pydantic==2.5.2
    # via
    #   langchain
    #   langchain-core
    #   langsmith
    #   mistralai
pydantic-core==2.14.5
    # via pydantic
pyproject-hooks==1.0.0
    # via build
python-dateutil==2.8.2
    # via
    #   botocore
    #   google-cloud-bigquery
    #   pandas
python-dotenv==1.0.0
    # via -r requirements.in
pytz==2024.1
    # via pandas
pyyaml==6.0.1
    # via
    #   huggingface-hub
    #   langchain
    #   langchain-community
    #   langchain-core
regex==2023.10.3
    # via tiktoken
requests==2.31.0
    # via
    #   cohere
    #   google-api-core
    #   google-cloud-bigquery
    #   google-cloud-storage
    #   huggingface-hub
    #   langchain
    #   langchain-community
    #   langchain-core
    #   langsmith
    #   openai
    #   tiktoken
rsa==4.9
    # via google-auth
s3transfer==0.7.0
    # via boto3
shapely==2.0.2
    # via google-cloud-aiplatform
six==1.16.0
    # via python-dateutil
sniffio==1.3.0
    # via
    #   anyio
    #   httpx
sqlalchemy==2.0.20
    # via
    #   langchain
    #   langchain-community
tenacity==8.2.3
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
tiktoken==0.5.1
    # via -r requirements.in
tokenizers==0.15.2
    # via langchain-mistralai
tqdm==4.66.1
    # via
    #   huggingface-hub
    #   openai
typing-extensions==4.7.1
    # via
    #   huggingface-hub
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typing-inspect
typing-inspect==0.9.0
    # via dataclasses-json
tzdata==2024.1
    # via pandas
urllib3==1.26.16
    # via
    #   botocore
    #   cohere
    #   elastic-transport
    #   requests
werkzeug==2.3.7
    # via flask
wheel==0.41.3
    # via pip-tools
yarl==1.9.2
    # via aiohttp
zipp==3.17.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
pypdf==5.1.0
langchain-text-splitters==0.0.2